changed abstractArticle [done], AbstractFormPage [WIP], AbstractSearchPage [WIP]
changed the service method to it's own controller
