+++ src/Kunstmaan/NodeSearchBundle/Entity/AbstractSearchPage.php
@@ -2,90 +2,24 @@
 
 namespace Kunstmaan\NodeSearchBundle\Entity;
 
+use Kunstmaan\NodeBundle\Controller\SlugActionInterface;
 use Kunstmaan\NodeBundle\Entity\AbstractPage;
-use Kunstmaan\NodeBundle\Helper\RenderContext;
-use Kunstmaan\NodeSearchBundle\PagerFanta\Adapter\SearcherRequestAdapter;
-use Kunstmaan\NodeSearchBundle\Search\AbstractElasticaSearcher;
 use Kunstmaan\SearchBundle\Helper\IndexableInterface;
-use Pagerfanta\Exception\NotValidCurrentPageException;
-use Pagerfanta\Pagerfanta;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+
 
 /**
  * AbstractSearchPage, extend this class to create your own SearchPage and extend the standard functionality
  */
-class AbstractSearchPage extends AbstractPage implements IndexableInterface
+class AbstractSearchPage extends AbstractPage implements IndexableInterface, SlugActionInterface
 {
-    /**
-     * Default number of search results to show per page (default: 10)
-     * @var int
-     */
-    private $defaultPerPage = 10;
+
 
     /**
-     * @param ContainerInterface $container
-     * @param Request            $request
-     * @param RenderContext      $context
-     *
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse|void
+     * @return string
      */
-    public function service(ContainerInterface $container, Request $request, RenderContext $context)
+    public function getControllerAction()
     {
-        // Perform a search if there is a queryString available
-        if ($request->query->has('query')) {
-            $pagerfanta            = $this->search($container, $request, $context);
-            $context['pagerfanta'] = $pagerfanta;
-        }
-    }
-
-    /**
-     * @param ContainerInterface $container
-     * @param Request            $request
-     *
-     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
-     * @return Pagerfanta
-     */
-    public function search(ContainerInterface $container, Request $request, RenderContext $context)
-    {
-        // Retrieve the current page number from the URL, if not present of lower than 1, set it to 1
-        $pageNumber = $this->getRequestedPage($request);
-        $searcher   = $container->get($this->getSearcher());
-        $this->applySearchParams($searcher, $request, $context);
-
-        $adapter    = new SearcherRequestAdapter($searcher);
-        $pagerfanta = new Pagerfanta($adapter);
-        try {
-            $pagerfanta
-                ->setMaxPerPage($this->getDefaultPerPage())
-                ->setCurrentPage($pageNumber);
-        } catch (NotValidCurrentPageException $e) {
-            throw new NotFoundHttpException();
-        }
-
-        return $pagerfanta;
-    }
-
-    /**
-     * @param AbstractElasticaSearcher $searcher
-     * @param Request                  $request
-     * @param RenderContext            $context
-     */
-    protected function applySearchParams(AbstractElasticaSearcher $searcher, Request $request, RenderContext $context)
-    {
-        // Retrieve the search parameters
-        $queryString = trim($request->query->get('query'));
-        $queryType   = $request->query->get('type');
-        $lang        = $request->getLocale();
-
-        $context['q_query'] = $queryString;
-        $context['q_type']  = $queryType;
-
-        $searcher
-            ->setData($this->sanitizeSearchQuery($queryString))
-            ->setContentType($queryType)
-            ->setLanguage($lang);
+        return 'KunstmaanNodeSearchBundle:AbstractSearchPage:service';
     }
 
     /**
@@ -115,55 +49,8 @@ class AbstractSearchPage extends AbstractPage implements IndexableInterface
     /**
      * @return string
      */
-    protected function getSearcher()
+    public function getSearcher()
     {
         return 'kunstmaan_node_search.search.node';
-    }
-
-    /**
-     * @param int $defaultPerPage
-     *
-     * @return AbstractSearchPage
-     */
-    public function setDefaultPerPage($defaultPerPage)
-    {
-        $this->defaultPerPage = $defaultPerPage;
-
-        return $this;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDefaultPerPage()
-    {
-        return $this->defaultPerPage;
-    }
-
-    /**
-     * Currently we just search for a complete match...
-     *
-     * @param string $query
-     *
-     * @return string
-     */
-    protected function sanitizeSearchQuery($query)
-    {
-        return '"' . $query . '"';
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return int
-     */
-    private function getRequestedPage(Request $request)
-    {
-        $pageNumber = $request->query->getInt('page', 1);
-        if (!$pageNumber || $pageNumber < 1) {
-            $pageNumber = 1;
-        }
-
-        return $pageNumber;
     }
 }
